Add JavaScript compilation step
**Why**: For improved maintainability as we seek to expand JavaScript implementation.
