+++ .eslintignore
@@ -1,1 +1,2 @@
 *.min.js
+assets/js/build
