+++ package.json
@@ -8,12 +8,24 @@
   },
   "devDependencies": {
     "@18f/eslint-plugin-identity": "^2.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.16.0",
+    "@typescript-eslint/parser": "^5.16.0",
+    "esbuild": "^0.14.27",
     "eslint": "^8.11.0",
     "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "prettier": "^2.5.1"
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "prettier": "^2.5.1",
+    "typescript": "^4.6.2"
   },
   "scripts": {
+    "clean": "rm -rf assets/js/build",
+    "prebuild": "npm run clean",
+    "build": "esbuild `find assets/js -type f -not -name '*.min.*'` --outdir=assets/js/build --bundle --format=esm",
+    "prestart": "npm run clean",
+    "start": "npm run build -- --watch",
     "lint": "eslint assets/js --ext .js,.jsx,.ts,.tsx"
   }
 }
